+++ endpoints/publicProxy/interstitialUi/embed.go
@@ -1,0 +1,6 @@
+package interstitialUi
+
+import "embed"
+
+//go:embed index.html
+var FS embed.FS